# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiErrorMessage {
  error: string;
}

export interface Usuario {
  id: number;
  name: string;
  email: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthSession {
  token: string;
  user: Usuario;
}

export interface NotaCategoria {
  /**
     * @minimum 1
     * @maximum 5
     */
  nota: number;
  comentario?: string;
}

export interface AvaliacaoNotas {
  apartamento?: NotaCategoria;
  condominio?: NotaCategoria;
  proprietario?: NotaCategoria;
  imobiliaria?: NotaCategoria;
}

export interface Avaliacao {
  id: number;
  imovelId: number;
  userId: number;
  userName: string;
  data: string;
  notas: AvaliacaoNotas;
  oQueGostariaDeSaber: string;
}

export interface AvaliacaoInput {
  imovelId: number;
  notas: AvaliacaoNotas;
  /** @minLength 20 */
  oQueGostariaDeSaber: string;
}

export interface Imovel {
  id: number;
  nome: string;
  endereco: string;
  bairro: string;
  cidade: string;
  avaliacoes: Avaliacao[];
}

export type ListAvaliacoesParams = {
imovelId?: number;
};