# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /imoveis/{id}:
    get:
      operationId: getImovel
      tags: [imoveis]
      summary: Get a property with its reviews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Imovel"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
...
components:
  schemas:
...
    ApiErrorMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    NotaCategoria:
      type: object
      properties:
        nota:
          type: number
          minimum: 1
          maximum: 5
        comentario:
          type: string
      required: [nota]
    AvaliacaoNotas:
      type: object
      properties:
        apartamento:
          $ref: "#/components/schemas/NotaCategoria"
        condominio:
          $ref: "#/components/schemas/NotaCategoria"
        proprietario:
          $ref: "#/components/schemas/NotaCategoria"
        imobiliaria:
          $ref: "#/components/schemas/NotaCategoria"
    Avaliacao:
      type: object
      properties:
        id:
          type: number
        imovelId:
          type: number
        userId:
          type: number
        userName:
          type: string
        data:
          type: string
          format: date-time
        notas:
          $ref: "#/components/schemas/AvaliacaoNotas"
        oQueGostariaDeSaber:
          type: string
      required: [id, imovelId, userId, userName, data, notas, oQueGostariaDeSaber]
...
    Imovel:
      type: object
      properties:
        id:
          type: number
        nome:
          type: string
        endereco:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        avaliacoes:
          type: array
          items:
            $ref: "#/components/schemas/Avaliacao"
      required: [id, nome, endereco, bairro, cidade, avaliacoes]